add initial svm results

+++ project/py_scripts/svm.py
@@ -1,4 +1,6 @@
+import csv
 import numpy as np
+import os
 from random import shuffle
 import re
 import subprocess
@@ -12,13 +14,15 @@ w_out = open('../data/svm_out/w.out','w')
 w_out_sorted = open('../data/svm_out/w_sorted.out','w')
 # TODO better name for class
 class Instance():
-    def __init__(self, train_dat_num, test_dat_num, c=None,j=None, t=None):
+    def __init__(self, train_dat_num, val_dat_num, test_dat_num, c=None,j=None, t=None):
         self.train_data_num = train_dat_num
         self.test_data_num = test_dat_num
+        self.val_dat_num = val_dat_num
         self.c = c
         self.j = j
         self.t = t
         self.train_file = '../data/svm_norm/dat_' + str(train_dat_num) +'.txt'
+        self.val_file = '../data/svm_norm/dat_' + str(train_dat_num) +'.txt'
         self.test_file = '../data/svm_norm/dat_' + str(test_dat_num) +'.txt'
         prefix = '../data/svm_out/dat_' + str(train_dat_num)
         if c != None:
@@ -36,6 +40,9 @@ class Instance():
         self.np = None
         self.tp = None
         self.f1 = None
+        self.f_half = None
+        self.precision = None
+        self.recall = None
         self.w = None
 
 class Kernel():
@@ -52,6 +59,9 @@ class Kernel():
         self.b = float(threshold)
 
 def learn(instance):
+    if os.path.isfile(instance.model):
+        print "Skipped training " + instance.model
+        return
     args_list = ['./svm_light/svm_learn']
     if instance.c is not None:
         args_list.append('-c')
@@ -71,6 +81,9 @@ def learn(instance):
     f_tmp.close()
 
 def classify(instance):
+    if os.path.isfile(instance.prediction):
+        print "Skipping classifying " + instance.classify_out
+        return
     output = open(instance.classify_out, 'w')
     args = ('./svm_light/svm_classify', instance.test_file, instance.model, instance.prediction)
     popen = subprocess.Popen(args, stdout=output)
@@ -114,9 +127,10 @@ def find_results(instance):
     instance.fp = sum([1 for (t, p) in zip(y_data, predictions) if t <= 0 and p > 0])
     instance.tp = sum([1 for (t, p) in zip(y_data, predictions) if t > 0 and p > 0])
     instance.accuracy = 1 - float(instance.fn + instance.fp)/len(y_data)
-    precision = instance.tp / float(instance.tp+instance.fp)
-    recall = instance.tp / float(instance.tp + instance.fn)
-    instance.f1 = 2 * precision * recall / float(precision + recall)
+    instance.precision = instance.tp / float(instance.tp+instance.fp)
+    instance.recall = instance.tp / float(instance.tp + instance.fn)
+    instance.f1 = 2 * instance.precision * instance.recall / float(instance.precision + instance.recall)
+    instance.f_half = 1.25 * instance.precision * instance.recall / float(.25*instance.precision + instance.recall)
     f.close()
 
 def process(instance):
@@ -126,24 +140,35 @@ def process(instance):
     find_w(instance)
 
 def run(train_test_pairs, j_vals=[None], c_vals=[None], t=None):
+    output = open('../data/svm_out/output.csv', 'wb')
+    writer = csv.writer(output)
     table = texttable.Texttable()
-    table.header(['Train File', 'Test File', 'j', 'c',  '# FN ', '# FP', 'Accuracy','F1'])
-    for (train, test) in train_test_pairs:
+    header = ['Train File', 'Test File', 'j', 'c', 't',  '# FN ', '# FP', 'Accuracy','Precision','Recall','F1','F.5']
+    table.header(header)
+    writer.writerow(header)
+    for (train, val, test) in train_test_pairs:
         for j in j_vals:
             for c in c_vals:
-                instance = Instance(train, test, c=c, j=j, t=t)
+                instance = Instance(train, val, test, c=c, j=j, t=t)
                 process(instance)
-                table.add_row([instance.train_file, instance.test_file, instance.j, instance.c, instance.fn, instance.fp, instance.accuracy,instance.f1])
-                print instance.train_file, instance.test_file, instance.j, instance.c, instance.fn, instance.fp, instance.accuracy, instance.f1
+                row= [instance.train_file, instance.test_file, instance.j, instance.c, instance.t, instance.fn, instance.fp, instance.accuracy, instance.precision, instance.recall, instance.f1,instance.f_half]
+                table.add_row(row)
+                writer.writerow(row)
+                print row
     print "---------------------------"
     print
     print table.draw()
 
+
+
 DAT_NUMS = range(1,11) #TODO zero index the data files
 
-J_VALS = [.75,1]
-C_VALS = [0.1,1.0,10.0]
-pairs = [(i,10) for i in DAT_NUMS[:-1]]
+J_VALS = [.6,.75,.9]
+C_VALS = [1,10,100]
 
-run(pairs,j_vals=J_VALS,t=2)
+#pairs = [(1,9,10)]
+pairs = [(i,9,10) for i in DAT_NUMS[:-2]]
+
+
+run(pairs,c_vals=C_VALS,j_vals=J_VALS,t=2)
